+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <script src="./dist/app.js"></script>
+    <script type="module" src="./dist/app.js"></script>
 </head>
 <body>
