+++ src/classes/CartItem.ts
@@ -1,0 +1,9 @@
+import Book from "./book";
+
+export default interface CartItem {
+    book: Book;
+    quantity: number;
+    price: number;
+    subtotal: number;
+    discountVoucher: string | number ;
+  }
