+++ src/classes/creditCard.ts
@@ -1,0 +1,11 @@
+import Payment from "./payment";
+
+class CreditCardPayment extends Payment {
+    constructor(amount: number, private creditCardNumber: string) {
+        super(amount);
+        this.creditCardNumber = creditCardNumber;
+    }
+    processPayment(): void {
+        console.log(`Credit card payment of $${this.amount} with card number ${this.creditCardNumber}`);
+    }
+}
