+++ src/classes/payment.ts
@@ -1,0 +1,6 @@
+export default abstract class Payment {
+    constructor(protected amount: number) {
+        this.amount = amount;
+    }
+    abstract processPayment(): void ;
+}
