+++ src/classes/paypal.ts
@@ -1,0 +1,11 @@
+import Payment from "./payment";
+
+class PayPalPayment extends Payment {
+    constructor(amount: number, private email: string) {
+        super(amount);
+        this.email = email;
+    }
+    processPayment(): void {
+        console.log(`PayPal payment of $${this.amount} with email ${this.email}`);
+    }
+}
