+++ src/classes/shoppingCart.ts
@@ -1,0 +1,7 @@
+import Book from "./book";
+
+export default interface ShoppingCart {
+    addBook(book: Book): void;
+    removeBook(book: Book): void;
+    viewCart(): Book[];
+}
